// src/app/models/cupones.ts
export class cupones {
    idcupon: number;
    categoria: string;
    cupon: string;
    descripcion: string;
    tipocupon: string;
    limiteuso: number;
    numerouso: number;
    fechavalido: string;
    descuento: number;
    iduseradd: number;
    dateadd: string;
    idusermod: number;
    datemod: string;
}